requirements.txt
annotated-types==0.7.0
anyio==3.7.1
atproto==0.0.61
beautifulsoup4==4.12.3
blurhash==1.1.4
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.7
cryptography==42.0.8
decorator==5.1.1
dnspython==2.7.0
exceptiongroup==1.2.2
feedparser==6.0.10
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==2.10
importlib-metadata==6.7.0
libipld==3.0.1
Mastodon.py==2.0.1
oauthlib==3.1.0
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.24.0
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
sniffio==1.3.1
soupsieve==2.6
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.25.10
websockets==13.1
zipp==3.15.0
